Relocate `Specs` next to class declaration.
